added fuzz tests for FromBinary family of functions, moved to the codec_tests.go file
